Updates to prototype javascript loading issue
When upgrading the prototype some of the old code was conflicting with the new. This meant the javascript was not loading. Changes have been made to rectify this. Also the footer has been changed so that we can now directly manage the prototype from the link that use to be version of the prototype. Prototype has also been updated with the latest version